// app/types/novapost.ts
export type WareHouseDTO = {
    SiteKey: string;
    Description: string;
    SettlementDescription: string;
    SettlementAreaDescription: string;
};
export type SettlementAreaDTO = {
    Ref: string;
    Description: string;
};
export type SettlementDTO = {
    Ref: string;
    Description: string;
};